whew, at last seemingly good
