+++ TUF/TrustedMetadata.cs
@@ -1,279 +1,533 @@
 using System.Diagnostics.CodeAnalysis;
 
-using TUF.Models;
-using TUF.Models.Primitives;
-using TUF.Models.Roles.Root;
-using TUF.Models.Roles.Snapshot;
-using TUF.Models.Roles.Targets;
-using TUF.Models.Roles.Timestamp;
-using TUF.Serialization;
+using Serde;
 
-namespace TUF;
+namespace TUF.Models;
 
-public class TrustedMetadata(RootMetadata root)
+/// <summary>
+/// Represents trusted TUF metadata in various states of completion.
+/// Provides secure, progressive loading of metadata following TUF specification order and verification requirements.
+/// </summary>
+/// <remarks>
+/// TUF metadata must be loaded and verified in a specific order to maintain security:
+/// 1. Root metadata (establishes trust)
+/// 2. Timestamp metadata (provides freshness)
+/// 3. Snapshot metadata (provides consistency)
+/// 4. Targets metadata (provides file catalog)
+/// 
+/// Each stage validates the next using cryptographic signatures and role delegation,
+/// ensuring the entire metadata chain is trustworthy.
+/// </remarks>
+public class TrustedMetadata
 {
-    public RootMetadata Root { get; private set; } = root;
-    public required DateTimeOffset RefTime { get; init; }
+    /// <summary>
+    /// Root metadata containing the foundation of trust for the repository.
+    /// Contains key assignments and role definitions for all other metadata types.
+    /// </summary>
+    public Metadata<Root> Root { get; private set; }
 
+    /// <summary>
+    /// Reference time used for expiration checks.
+    /// All metadata expiration times are evaluated against this timestamp.
+    /// </summary>
+    public DateTimeOffset RefTime { get; init; }
+
+    /// <summary>
+    /// Creates a new trusted metadata instance starting with root metadata.
+    /// </summary>
+    /// <param name="root">Root metadata that has been independently verified</param>
+    /// <param name="refTime">Reference time for expiration checks</param>
+    protected TrustedMetadata(Metadata<Root> root, DateTimeOffset refTime)
+    {
+        Root = root;
+        RefTime = refTime;
+    }
+
+    /// <summary>
+    /// Creates trusted metadata from root metadata bytes.
+    /// This is the entry point for establishing trust in a TUF repository.
+    /// </summary>
+    /// <param name="rootData">JSON bytes of root metadata</param>
+    /// <returns>Trusted metadata initialized with verified root</returns>
+    /// <exception cref="Exception">Thrown if root metadata is invalid or verification fails</exception>
+    /// <remarks>
+    /// Root metadata is special because it is self-signed - it must be independently
+    /// verified through out-of-band means (secure distribution, key fingerprint verification, etc.)
+    /// before being trusted.
+    /// </remarks>
     public static TrustedMetadata CreateFromRootData(byte[] rootData)
     {
         var refTime = DateTimeOffset.UtcNow;
-        var newRoot = MetadataSerializer.Deserialize<RootMetadata>(rootData);
-        if (newRoot is null)
+
+        // Deserialize root metadata
+        var rootMetadata = Serde.Json.JsonSerializer.Deserialize<Metadata<Root>, MetadataProxy.De<Root>>(rootData, MetadataProxy.De<Root>.Instance);
+        if (rootMetadata == null)
         {
             throw new Exception("Failed to deserialize root metadata");
         }
-        MetadataExtensions.VerifyRootRole<RootMetadata, RootMetadata, Root>(newRoot, Models.Roles.Root.Root.TypeLabel, newRoot);
-        return new TrustedMetadata(newRoot)
+
+        // Root metadata is self-signed, so we verify it against itself
+        rootMetadata.VerifyRole("root", rootMetadata);
+
+        // Check if root is expired
+        if (IsExpired(rootMetadata.Signed.Expires, refTime))
         {
-            RefTime = refTime
-        };
+            throw new Exception("Root metadata is expired");
+        }
+
+        return new TrustedMetadata(rootMetadata, refTime);
     }
 
-    public TrustedMetadata UpdateRoot(byte[] rootData)
+    /// <summary>
+    /// Updates root metadata to a new version.
+    /// Verifies the new root using the current root's trust relationships.
+    /// </summary>
+    /// <param name="rootData">JSON bytes of new root metadata</param>
+    /// <returns>This trusted metadata instance with updated root</returns>
+    /// <exception cref="Exception">Thrown if verification fails or version is invalid</exception>
+    /// <remarks>
+    /// Root updates require careful verification:
+    /// 1. New root must be signed by keys authorized in current root
+    /// 2. Version must increment by exactly 1
+    /// 3. New root must self-verify (ensuring key rotation is valid)
+    /// </remarks>
+    public virtual TrustedMetadata UpdateRoot(byte[] rootData)
     {
-        var newRoot = MetadataSerializer.Deserialize<RootMetadata>(rootData);
-        if (newRoot is null)
+        var newRoot = Serde.Json.JsonSerializer.Deserialize<Metadata<Root>, MetadataProxy.De<Root>>(rootData, MetadataProxy.De<Root>.Instance);
+        if (newRoot == null)
         {
-            throw new Exception("Failed to deserialize root metadata");
+            throw new Exception("Failed to deserialize new root metadata");
         }
-        MetadataExtensions.VerifyRootRole<RootMetadata, RootMetadata, Root>(Root, Models.Roles.Root.Root.TypeLabel, newRoot);
+
+        // Verify new root against current root's delegation
+        Root.VerifyRole("root", newRoot);
+
+        // Verify version increment (must be exactly +1 for root)
         if (newRoot.Signed.Version != Root.Signed.Version + 1)
         {
-            throw new Exception($"Invalid root metadata version: {newRoot.Signed.Version}");
+            throw new Exception($"Invalid root version: expected {Root.Signed.Version + 1}, got {newRoot.Signed.Version}");
         }
-        MetadataExtensions.VerifyRootRole<RootMetadata, RootMetadata, Root>(newRoot, Models.Roles.Root.Root.TypeLabel, newRoot);
+
+        // New root must self-verify to ensure key rotation is valid
+        newRoot.VerifyRole("root", newRoot);
+
+        // Check expiration
+        if (IsExpired(newRoot.Signed.Expires, RefTime))
+        {
+            throw new Exception("New root metadata is expired");
+        }
+
         Root = newRoot;
         return this;
     }
 
-    protected RootAndTimestampTrustedMetadata UpdateTimestampInternal(byte[] timestampData)
+    /// <summary>
+    /// Updates timestamp metadata. This advances the trusted metadata to include timestamp information.
+    /// </summary>
+    /// <param name="timestampData">JSON bytes of timestamp metadata</param>
+    /// <returns>New trusted metadata instance including timestamp</returns>
+    /// <exception cref="Exception">Thrown if verification fails or root is expired</exception>
+    public virtual TrustedMetadataWithTimestamp UpdateTimestamp(byte[] timestampData)
     {
-        if (Root.IsExpired(RefTime))
+        // Root must not be expired before we can update timestamp
+        if (IsExpired(Root.Signed.Expires, RefTime))
         {
             throw new Exception("Cannot update timestamp with expired root metadata");
         }
-        var newTimestamp = MetadataSerializer.Deserialize<TimestampMetadata>(timestampData);
-        if (newTimestamp is null)
+
+        var timestamp = Serde.Json.JsonSerializer.Deserialize<Metadata<Timestamp>, MetadataProxy.De<Timestamp>>(timestampData, MetadataProxy.De<Timestamp>.Instance);
+        if (timestamp == null)
         {
             throw new Exception("Failed to deserialize timestamp metadata");
         }
-        MetadataExtensions.VerifyRootRole<RootMetadata, TimestampMetadata, Timestamp>(Root, Models.Roles.Timestamp.Timestamp.TypeLabel, newTimestamp);
-        var newTrustedMetadata = new RootAndTimestampTrustedMetadata(Root, newTimestamp)
+
+        // Verify timestamp against root's timestamp role
+        Root.VerifyRole("timestamp", timestamp);
+
+        // Check expiration
+        if (IsExpired(timestamp.Signed.Expires, RefTime))
         {
-            RefTime = RefTime
-        };
-        return newTrustedMetadata;
+            throw new Exception("Timestamp metadata is expired");
+        }
+
+        return new TrustedMetadataWithTimestamp(Root, timestamp, RefTime);
     }
 
-    public virtual RootAndTimestampTrustedMetadata UpdateTimeStamp(byte[] timestampData)
+    /// <summary>
+    /// Helper method to check if metadata has expired.
+    /// </summary>
+    /// <param name="expiresString">ISO 8601 expiration timestamp</param>
+    /// <param name="refTime">Reference time to compare against</param>
+    /// <returns>True if expired, false otherwise</returns>
+    protected static bool IsExpired(string expiresString, DateTimeOffset refTime)
     {
-        var newTrustedMetadata = UpdateTimestampInternal(timestampData);
-        newTrustedMetadata.CheckFinalTimestamp();
-        return newTrustedMetadata;
+        if (!DateTimeOffset.TryParse(expiresString, out var expires))
+        {
+            throw new Exception($"Invalid expiration time format: {expiresString}");
+        }
+        return refTime >= expires;
     }
 
-    public virtual TrustedMetadata UpdateSnapshot(byte[] snapshotData, bool isTrusted)
+    // Virtual methods that are only valid in later stages
+    public virtual TrustedMetadataWithTimestamp UpdateSnapshot(byte[] snapshotData, bool isTrusted)
     {
-        throw new NotImplementedException("Cannot update snapshot before timestamp");
+        throw new InvalidOperationException("Cannot update snapshot before timestamp");
     }
-    public virtual TrustedMetadata UpdateTargets(byte[] targetsData)
+
+    public virtual TrustedMetadataWithTimestamp UpdateDelegatedTargets(byte[] delegatedTargetsData, string roleName, string delegatorRoleName)
     {
-        return UpdateDelegatedTargets(targetsData, Models.Roles.Targets.TargetsRole.TypeLabel, Models.Roles.Root.Root.TypeLabel);
-    }
-    public virtual TrustedMetadata UpdateDelegatedTargets(byte[] delegatedTargetsData, string roleName, string delegatorRoleName)
-    {
-        throw new NotImplementedException("Cannot load targets before snapshot");
+        throw new InvalidOperationException("Cannot update delegated targets before snapshot");
     }
 }
 
+/// <summary>
+/// Trusted metadata that includes timestamp information.
+/// Provides snapshot file metadata and enables loading snapshot metadata.
+/// </summary>
 [method: SetsRequiredMembers]
-public class RootAndTimestampTrustedMetadata(RootMetadata root, TimestampMetadata timestamp) : TrustedMetadata(root)
+public class TrustedMetadataWithTimestamp(Metadata<Root> root, Metadata<Timestamp> timestamp, DateTimeOffset refTime)
+    : TrustedMetadata(root, refTime)
 {
-    public TimestampMetadata Timestamp { get; private set; } = timestamp;
+    /// <summary>
+    /// Timestamp metadata providing freshness guarantees and snapshot file metadata.
+    /// </summary>
+    public Metadata<Timestamp> Timestamp { get; private set; } = timestamp;
 
-    protected FileMetadata SnapshotMeta => Timestamp.Signed.SnapshotFileMetadata;
+    /// <summary>
+    /// Gets the snapshot file metadata from the timestamp.
+    /// This describes the expected snapshot file (version, length, hashes).
+    /// </summary>
+    protected FileMetadata SnapshotMeta => Timestamp.Signed.Meta["snapshot.json"];
 
-    public override RootAndTimestampTrustedMetadata UpdateTimeStamp(byte[] timestampData)
+    /// <summary>
+    /// Updates timestamp metadata with version and freshness validation.
+    /// </summary>
+    /// <param name="timestampData">JSON bytes of new timestamp metadata</param>
+    /// <returns>New or existing trusted metadata instance</returns>
+    /// <exception cref="Exception">Thrown if verification fails or versions are invalid</exception>
+    public override TrustedMetadataWithTimestamp UpdateTimestamp(byte[] timestampData)
     {
-        var newTrustedMetadata = UpdateTimestampInternal(timestampData);
-        // check that new timestamp isn't older than current timestamp
-        if (newTrustedMetadata.Timestamp.Signed.Version < Timestamp.Signed.Version)
+        if (IsExpired(Root.Signed.Expires, RefTime))
+        {
+            throw new Exception("Cannot update timestamp with expired root metadata");
+        }
+
+        var newTimestamp = Serde.Json.JsonSerializer.Deserialize<Metadata<Timestamp>, MetadataProxy.De<Timestamp>>(timestampData, MetadataProxy.De<Timestamp>.Instance);
+        if (newTimestamp == null)
+        {
+            throw new Exception("Failed to deserialize new timestamp metadata");
+        }
+
+        // Verify against root
+        Root.VerifyRole("timestamp", newTimestamp);
+
+        // Check version progression
+        if (newTimestamp.Signed.Version < Timestamp.Signed.Version)
         {
             throw new Exception("New timestamp version is older than current timestamp version");
         }
-        // keep using existing if versions are equal
-        else if (newTrustedMetadata.Timestamp.Signed.Version == Timestamp.Signed.Version)
+
+        // If same version, keep existing
+        if (newTimestamp.Signed.Version == Timestamp.Signed.Version)
         {
             return this;
         }
-        // prevent rolling back snapshot version
-        if (newTrustedMetadata.SnapshotMeta.Version < SnapshotMeta.Version)
+
+        // Check expiration
+        if (IsExpired(newTimestamp.Signed.Expires, RefTime))
+        {
+            throw new Exception("New timestamp metadata is expired");
+        }
+
+        // Prevent snapshot rollback
+        var newSnapshotMeta = newTimestamp.Signed.Meta["snapshot.json"];
+        if (newSnapshotMeta.Version < SnapshotMeta.Version)
         {
             throw new Exception("New snapshot version is older than current snapshot version");
         }
-        newTrustedMetadata.CheckFinalTimestamp();
-        return newTrustedMetadata;
+
+        return new TrustedMetadataWithTimestamp(Root, newTimestamp, RefTime);
     }
 
-    protected RootAndTimestampAndSnapshotTrustedMetadata UpdateSnapshotInternal(byte[] snapshotData, bool isTrusted)
+    /// <summary>
+    /// Updates snapshot metadata. Advances to the next stage of trusted metadata.
+    /// </summary>
+    /// <param name="snapshotData">JSON bytes of snapshot metadata</param>
+    /// <param name="isTrusted">Whether the data comes from a trusted source (bypasses length/hash verification)</param>
+    /// <returns>Trusted metadata including snapshot information</returns>
+    public override TrustedMetadataWithSnapshot UpdateSnapshot(byte[] snapshotData, bool isTrusted)
     {
-        CheckFinalTimestamp();
+        // Verify integrity if not from trusted source
         if (!isTrusted)
         {
-            SnapshotMeta.VerifyLengthHashes(snapshotData);
+            VerifyFileIntegrity(snapshotData, SnapshotMeta);
         }
-        var newSnapshot = MetadataSerializer.Deserialize<SnapshotMetadata>(snapshotData);
-        if (newSnapshot is null)
+
+        var snapshot = Serde.Json.JsonSerializer.Deserialize<Metadata<Snapshot>, MetadataProxy.De<Snapshot>>(snapshotData, MetadataProxy.De<Snapshot>.Instance);
+        if (snapshot == null)
         {
             throw new Exception("Failed to deserialize snapshot metadata");
         }
-        MetadataExtensions.VerifyRootRole<RootMetadata, SnapshotMetadata, Snapshot>(Root, Snapshot.TypeLabel, newSnapshot);
-        var newThing = new RootAndTimestampAndSnapshotTrustedMetadata(Root, Timestamp, newSnapshot)
+
+        // Verify against root's snapshot role
+        Root.VerifyRole("snapshot", snapshot);
+
+        // Version must match what timestamp expects
+        if (snapshot.Signed.Version != SnapshotMeta.Version)
         {
-            RefTime = RefTime
-        };
-        return newThing;
+            throw new Exception("Snapshot version doesn't match timestamp metadata");
+        }
+
+        // Check expiration
+        if (IsExpired(snapshot.Signed.Expires, RefTime))
+        {
+            throw new Exception("Snapshot metadata is expired");
+        }
+
+        return new TrustedMetadataWithSnapshot(Root, Timestamp, snapshot, RefTime);
     }
 
-    public override RootAndTimestampAndSnapshotTrustedMetadata UpdateSnapshot(byte[] snapshotData, bool isTrusted)
+    /// <summary>
+    /// Verifies file integrity using length and cryptographic hashes.
+    /// </summary>
+    /// <param name="data">File data to verify</param>
+    /// <param name="fileMeta">Expected file metadata (length, hashes)</param>
+    /// <exception cref="Exception">Thrown if verification fails</exception>
+    public static void VerifyFileIntegrity(byte[] data, FileMetadata fileMeta)
     {
-        var newThing = UpdateSnapshotInternal(snapshotData, isTrusted);
-        newThing.CheckFinalSnapshot();
-        return newThing;
-    }
-
-    internal void CheckFinalTimestamp()
-    {
-        if (Timestamp.IsExpired(RefTime))
+        // Verify length
+        if (data.Length != fileMeta.Length)
         {
-            throw new Exception("Final timestamp is expired");
+            throw new Exception($"File length mismatch: expected {fileMeta.Length}, got {data.Length}");
+        }
+
+        // Verify at least one hash
+        bool hashVerified = false;
+        if (fileMeta.Hashes is null)
+        {
+            throw new Exception("File metadata is missing hashes");
+        }
+        foreach (var (algorithm, expectedHash) in fileMeta.Hashes)
+        {
+            string? actualHash = algorithm.ToLowerInvariant() switch
+            {
+                "sha256" => Convert.ToHexString(System.Security.Cryptography.SHA256.HashData(data)).ToLowerInvariant(),
+                "sha512" => Convert.ToHexString(System.Security.Cryptography.SHA512.HashData(data)).ToLowerInvariant(),
+                _ => null // Skip unsupported hash algorithms
+            };
+
+            if (actualHash != null && actualHash == expectedHash.ToLowerInvariant())
+            {
+                hashVerified = true;
+                break;
+            }
+        }
+
+        if (!hashVerified)
+        {
+            throw new Exception("No hash verification succeeded");
         }
     }
 }
 
+/// <summary>
+/// Trusted metadata that includes snapshot information.
+/// Provides consistent view of all target metadata and enables loading target files.
+/// </summary>
 [method: SetsRequiredMembers]
-public class RootAndTimestampAndSnapshotTrustedMetadata(RootMetadata root, TimestampMetadata timestamp, SnapshotMetadata snapshot) : RootAndTimestampTrustedMetadata(root, timestamp)
+public class TrustedMetadataWithSnapshot(Metadata<Root> root, Metadata<Timestamp> timestamp, Metadata<Snapshot> snapshot, DateTimeOffset refTime)
+    : TrustedMetadataWithTimestamp(root, timestamp, refTime)
 {
-    public SnapshotMetadata Snapshot { get; private set; } = snapshot;
+    /// <summary>
+    /// Snapshot metadata providing a consistent view of all available targets metadata.
+    /// </summary>
+    public Metadata<Snapshot> Snapshot { get; private set; } = snapshot;
 
-    public override RootAndTimestampAndSnapshotTrustedMetadata UpdateSnapshot(byte[] snapshotData, bool isTrusted)
+    /// <summary>
+    /// Updates snapshot metadata with consistency validation.
+    /// </summary>
+    /// <param name="snapshotData">JSON bytes of new snapshot metadata</param>
+    /// <param name="isTrusted">Whether the data comes from a trusted source</param>
+    /// <returns>New trusted metadata instance with updated snapshot</returns>
+    public override TrustedMetadataWithSnapshot UpdateSnapshot(byte[] snapshotData, bool isTrusted)
     {
-        var newThing = UpdateSnapshotInternal(snapshotData, isTrusted);
-        foreach (var fileEntry in Snapshot.Signed.Meta)
+        var newTrustedMetadata = (TrustedMetadataWithSnapshot)base.UpdateSnapshot(snapshotData, isTrusted);
+
+        // Ensure no target metadata versions go backwards
+        foreach (var (fileName, currentMeta) in Snapshot.Signed.Meta)
         {
-            if (!newThing.Snapshot.Signed.Meta.TryGetValue(fileEntry.Key, out var newMeta))
+            if (newTrustedMetadata.Snapshot.Signed.Meta.TryGetValue(fileName, out var newMeta))
             {
-                throw new Exception($"New snapshot metadata for {fileEntry.Key} is missing");
+                if (newMeta.Version < currentMeta.Version)
+                {
+                    throw new Exception($"New snapshot version for {fileName} is older than current version");
+                }
             }
-            if (newMeta.Version < fileEntry.Value.Version)
+            else
             {
-                throw new Exception($"New snapshot version for {fileEntry.Key} is older than current snapshot version");
+                throw new Exception($"New snapshot is missing metadata for {fileName}");
             }
         }
-        newThing.CheckFinalSnapshot();
-        return newThing;
+
+        return newTrustedMetadata;
     }
 
-    protected CoreTrustedMetadata UpdateDelegatedTargetsInternal(byte[] delegatedTargetsData, string roleName,
-        Action verifyDelegator,
-        Action<TargetsMetadata> verify,
-        Func<TargetsMetadata, CoreTrustedMetadata> acceptVerified)
+    /// <summary>
+    /// Updates delegated targets metadata.
+    /// </summary>
+    /// <param name="delegatedTargetsData">JSON bytes of delegated targets metadata</param>
+    /// <param name="roleName">Name of the delegated role</param>
+    /// <param name="delegatorRoleName">Name of the role that delegated authority</param>
+    /// <returns>Complete trusted metadata including the delegated targets</returns>
+    public override CompleteTrustedMetadata UpdateDelegatedTargets(byte[] delegatedTargetsData, string roleName, string delegatorRoleName)
     {
-        CheckFinalSnapshot();
-        verifyDelegator.Invoke();
-        if (!Snapshot.Signed.Meta.TryGetValue(new RelativePath(roleName + ".json"), out var fileMeta))
+        // Find the file metadata in snapshot
+        var fileName = roleName + ".json";
+        if (!Snapshot.Signed.Meta.TryGetValue(fileName, out var fileMeta))
         {
             throw new Exception($"File metadata for {roleName} not found in snapshot");
         }
-        fileMeta.VerifyLengthHashes(delegatedTargetsData);
-        var newTargets = MetadataSerializer.Deserialize<TargetsMetadata>(delegatedTargetsData);
-        if (newTargets is null)
-        {
-            throw new Exception("Failed to deserialize delegated targets");
-        }
-        verify.Invoke(newTargets);
-        if (newTargets.Signed.Version != fileMeta.Version)
-        {
-            throw new Exception("Target metadata out of sync with snapshot metadata");
-        }
-        if (newTargets.IsExpired(RefTime))
-        {
-            throw new Exception("Target metadata is expired");
-        }
-        return acceptVerified.Invoke(newTargets);
-    }
 
-    public override CoreTrustedMetadata UpdateDelegatedTargets(byte[] delegatedTargetsData, string roleName, string delegatorRoleName)
-    {
-        return UpdateDelegatedTargetsInternal(delegatedTargetsData, roleName,
-            () =>
-            {
-                if (delegatorRoleName != Models.Roles.Root.Root.TypeLabel)
-                {
-                    throw new NotImplementedException("Cannot load targets before delegator");
-                }
-            },
-            (newTargets) => MetadataExtensions.VerifyRootRole<RootMetadata, TargetsMetadata, TargetsRole>(Root, delegatorRoleName, newTargets),
-            (newTargets) => new CoreTrustedMetadata(Root, Timestamp, Snapshot, newTargets)
-            {
-                RefTime = RefTime
-            });
-    }
+        // Verify file integrity
+        VerifyFileIntegrity(delegatedTargetsData, fileMeta);
 
-    public void CheckFinalSnapshot()
-    {
-        if (Snapshot.IsExpired(RefTime))
+        // Deserialize targets metadata
+        var targets = Serde.Json.JsonSerializer.Deserialize<Metadata<Targets>, MetadataProxy.De<Targets>>(delegatedTargetsData, MetadataProxy.De<Targets>.Instance);
+        if (targets == null)
         {
-            throw new Exception("Final snapshot is expired");
+            throw new Exception("Failed to deserialize targets metadata");
         }
-        if (Snapshot.Signed.Version != SnapshotMeta.Version)
+
+        // Verify signatures based on delegator
+        if (delegatorRoleName == "root")
         {
-            throw new Exception("Final snapshot version is out of sync with timestamp snapshot metadata");
+            // Verify against root's targets role
+            Root.VerifyRole("targets", targets);
         }
+        else
+        {
+            // This would require having the delegator targets metadata loaded
+            // For now, we'll throw as this requires more complex state management
+            throw new NotImplementedException("Delegated role verification not yet implemented");
+        }
+
+        // Version must match snapshot expectation
+        if (targets.Signed.Version != fileMeta.Version)
+        {
+            throw new Exception("Targets version doesn't match snapshot metadata");
+        }
+
+        // Check expiration
+        if (IsExpired(targets.Signed.Expires, RefTime))
+        {
+            throw new Exception("Targets metadata is expired");
+        }
+
+        // Create the complete trusted metadata
+        var targetsMap = new Dictionary<string, Metadata<Targets>> { [roleName] = targets };
+        return new CompleteTrustedMetadata(Root, Timestamp, Snapshot, targetsMap, RefTime);
     }
 }
 
+/// <summary>
+/// Complete trusted metadata containing all loaded TUF metadata.
+/// Provides access to target files and supports ongoing metadata updates.
+/// </summary>
 [method: SetsRequiredMembers]
-public class CoreTrustedMetadata(RootMetadata root, TimestampMetadata timestamp, SnapshotMetadata snapshot, TargetsMetadata initialTarget) : RootAndTimestampAndSnapshotTrustedMetadata(root, timestamp, snapshot)
+public class CompleteTrustedMetadata(
+    Metadata<Root> root,
+    Metadata<Timestamp> timestamp,
+    Metadata<Snapshot> snapshot,
+    Dictionary<string, Metadata<Targets>> targets,
+    DateTimeOffset refTime)
+    : TrustedMetadataWithSnapshot(root, timestamp, snapshot, refTime)
 {
-    public TargetsMap Targets { get; private set; } = new(initialTarget);
+    /// <summary>
+    /// Map of all loaded targets metadata by role name.
+    /// Always includes at least the top-level "targets" role.
+    /// </summary>
+    public IReadOnlyDictionary<string, Metadata<Targets>> Targets { get; } = targets;
 
-    public override CoreTrustedMetadata UpdateDelegatedTargets(byte[] delegatedTargetsData, string roleName, string delegatorRoleName)
+    /// <summary>
+    /// Gets the top-level targets metadata.
+    /// </summary>
+    public Metadata<Targets> TopLevelTargets => Targets["targets"];
+
+    /// <summary>
+    /// Updates delegated targets metadata and adds it to the loaded set.
+    /// </summary>
+    /// <param name="delegatedTargetsData">JSON bytes of delegated targets metadata</param>
+    /// <param name="roleName">Name of the delegated role</param>
+    /// <param name="delegatorRoleName">Name of the role that delegated authority</param>
+    /// <returns>This trusted metadata instance with the new targets added</returns>
+    public override CompleteTrustedMetadata UpdateDelegatedTargets(byte[] delegatedTargetsData, string roleName, string delegatorRoleName)
     {
-        return UpdateDelegatedTargetsInternal(delegatedTargetsData, roleName,
-            () =>
-            {
-                if (delegatorRoleName != Models.Roles.Root.Root.TypeLabel)
-                {
-                    // verify that we have a target of this name
-                    if (!Targets.TryGetValue(delegatorRoleName, out var delegatorTargets))
-                    {
-                        throw new Exception($"Delegator role {delegatorRoleName} not found in loaded targets");
-                    }
-                }
-            },
-            (newTargets) =>
-            {
-                if (delegatorRoleName == Models.Roles.Root.Root.TypeLabel)
-                {
-                    MetadataExtensions.VerifyRootRole<RootMetadata, TargetsMetadata, TargetsRole>(Root, delegatorRoleName, newTargets);
-                }
-                else
-                {
-                    Targets[delegatorRoleName].VerifyDelegatedRole(roleName, newTargets);
-                }
-            },
-            (newTargets) =>
-            {
-                Targets[roleName] = newTargets;
-                return this;
-            }
-            );
-    }
-}
+        // Find the file metadata in snapshot
+        var fileName = roleName + ".json";
+        if (!Snapshot.Signed.Meta.TryGetValue(fileName, out var fileMeta))
+        {
+            throw new Exception($"File metadata for {roleName} not found in snapshot");
+        }
 
-public class TargetsMap(TargetsMetadata topLevelMetadata) : Dictionary<string, TargetsMetadata>([new KeyValuePair<string, TargetsMetadata>(Models.Roles.Targets.TargetsRole.TypeLabel, topLevelMetadata)])
-{
-    public TargetsMetadata TopLevelTargets => topLevelMetadata;
+        // Verify file integrity
+        VerifyFileIntegrity(delegatedTargetsData, fileMeta);
+
+        // Deserialize targets metadata
+        var targets = Serde.Json.JsonSerializer.Deserialize<Metadata<Targets>, MetadataProxy.De<Targets>>(delegatedTargetsData, MetadataProxy.De<Targets>.Instance);
+        if (targets == null)
+        {
+            throw new Exception("Failed to deserialize targets metadata");
+        }
+
+        // Verify signatures based on delegator
+        if (delegatorRoleName == "root")
+        {
+            Root.VerifyRole("targets", targets);
+        }
+        else if (Targets.TryGetValue(delegatorRoleName, out var delegatorTargets))
+        {
+            delegatorTargets.VerifyDelegatedRole(roleName, targets);
+        }
+        else
+        {
+            throw new Exception($"Delegator role {delegatorRoleName} not loaded");
+        }
+
+        // Version must match snapshot expectation
+        if (targets.Signed.Version != fileMeta.Version)
+        {
+            throw new Exception("Targets version doesn't match snapshot metadata");
+        }
+
+        // Check expiration
+        if (IsExpired(targets.Signed.Expires, RefTime))
+        {
+            throw new Exception("Targets metadata is expired");
+        }
+
+        // Add to loaded targets
+        var updatedTargets = new Dictionary<string, Metadata<Targets>>(Targets)
+        {
+            [roleName] = targets
+        };
+
+        return new CompleteTrustedMetadata(Root, Timestamp, Snapshot, updatedTargets, RefTime);
+    }
+
+    /// <summary>
+    /// Finds target file information across all loaded targets metadata.
+    /// </summary>
+    /// <param name="targetPath">Path to the target file</param>
+    /// <returns>Target file information if found, null otherwise</returns>
+    public TargetFile? FindTarget(string targetPath)
+    {
+        foreach (var targetsMetadata in Targets.Values)
+        {
+            if (targetsMetadata.Signed.TargetMap.TryGetValue(targetPath, out var target))
+            {
+                return target;
+            }
+        }
+        return null;
+    }
 }
