+++ TUF/Updater.cs
@@ -1,7 +1,6 @@
 using System.Runtime.InteropServices;
 
 using TUF.Models;
-using TUF.Models.Primitives;
 
 namespace TUF;
 
@@ -11,8 +10,8 @@ public class UpdaterConfig
     public uint MaxDelegations { get; init; } = 32;
     public uint RootMaxLength { get; init; } = 512000;
     public uint TimestampMaxLength { get; init; } = 16384;
-    public uint SnapshotMaxLength { get; init; } = 2000000;
-    public uint TargetsMaxLength { get; init; } = 5000000;
+    public int SnapshotMaxLength { get; init; } = 2000000;
+    public int TargetsMaxLength { get; init; } = 5000000;
 
     public byte[] LocalTrustedRoot { get; init; }
     public required string LocalMetadataDir { get; init; }
@@ -42,42 +41,44 @@ public class Updater
         _trusted = TrustedMetadata.CreateFromRootData(config.LocalTrustedRoot);
     }
 
-    public async Task Refresh()
+    public async Task RefreshAsync()
     {
         await OnlineRefresh();
     }
 
-    public Dictionary<RelativePath, Models.Roles.Targets.TargetMetadata> GetTopLevelTargets()
+    public Dictionary<string, TargetFile> GetTopLevelTargets()
     {
-        if (_trusted is not CoreTrustedMetadata c)
+        if (_trusted is not CompleteTrustedMetadata c)
         {
             throw new InvalidOperationException("Trusted metadata not loaded");
         }
-        return c.Targets.TopLevelTargets.Signed.Targets;
+        return c.TopLevelTargets.Signed.TargetMap;
     }
     
     public TrustedMetadata GetTrustedMetadataSet() => _trusted;
 
-    public async Task<TUF.Models.Roles.Targets.TargetMetadata> GetTargetInfo(string targetPath)
+    public async Task<(string RemotePath, TargetFile File)> GetTargetInfo(string targetPath)
     {
-        if (_trusted is not CoreTrustedMetadata c)
+        if (_trusted is not CompleteTrustedMetadata c)
         {
-            await Refresh();
+            await RefreshAsync();
         }
 
-        return await preOrderDepthFirstWalk(targetPath);
+        var (metaPath, fileData) = await PreOrderDepthFirstWalk(targetPath);
+        return (metaPath, fileData);
     }
 
-    public async Task<(string FilePath, byte[] Data)> DownloadTarget(Models.Roles.Targets.TargetMetadata targetFile, string? destinationFilePath, Uri? baseUrl)
+    public async Task<(string FilePath, byte[] Data)> DownloadTarget(TargetFile targetFile, string targetPath, string? destinationFilePath, Uri? baseUrl)
     {
-        var localDestinationPath = destinationFilePath ?? GenerateTargetFilePath(targetFile);
+        var localDestinationPath = destinationFilePath ?? Path.Combine(_config.LocalTargetsDir, targetPath);
         var targetBaseUrl = baseUrl ?? _config.RemoteTargetsUrl;
-        var targetRemotePath = targetFile.Path.RelPath;
-        if (_trusted.Root.Signed.ConsistentSnapshot is true && _config.PrefixTargetsWithHash)
+        var targetRemotePath = targetPath;
+        
+        if (_trusted.Root.Signed.ConsistentSnapshot == true && _config.PrefixTargetsWithHash)
         {
-            var hash = targetFile.Hashes.FirstOrDefault()?.HexEncodedValue;
-            var targetFileDir = Path.GetDirectoryName(targetFile.Path.RelPath);
-            var targetFileName = Path.GetFileName(targetFile.Path.RelPath)!;
+            var hash = targetFile.Hashes.FirstOrDefault().Value;
+            var targetFileDir = Path.GetDirectoryName(targetPath);
+            var targetFileName = Path.GetFileName(targetPath)!;
             if (targetFileDir is not null)
             {
                 targetRemotePath = Path.Combine(targetFileDir, $"{hash}.{targetFileName}");
@@ -89,8 +90,8 @@ public class Updater
         }
 
         var downloadUri = new Uri(targetBaseUrl, targetRemotePath);
-        var data = await DownloadFile(downloadUri, targetFile.Length);
-        targetFile.VerifyLengthHashes(data);
+        var data = await DownloadFile(downloadUri, (uint)targetFile.Length);
+        VerifyTargetFile(targetFile, data);
 
         if (!_config.DisableLocalCache)
         {
@@ -103,18 +104,23 @@ public class Updater
         return (localDestinationPath, data);
     }
 
-    public async Task<(string FilePath, byte[] Data)?> FindCachedTarget(Models.Roles.Targets.TargetMetadata targetFile, string? filePath)
+    public async Task<(string FilePath, byte[] Data)?> FindCachedTarget(TargetFile targetFile, string targetPath, string? filePath)
     {
         if (_config.DisableLocalCache)
         {
             return null;
         }
 
-        var localPath = filePath ?? GenerateTargetFilePath(targetFile);
+        var localPath = filePath ?? Path.Combine(_config.LocalTargetsDir, targetPath);
+        if (!File.Exists(localPath))
+        {
+            return null;
+        }
+        
         var fileBytes = await File.ReadAllBytesAsync(localPath);
         try
         {
-            targetFile.VerifyLengthHashes(fileBytes);
+            VerifyTargetFile(targetFile, fileBytes);
         }
         catch
         {
@@ -124,12 +130,47 @@ public class Updater
         return (localPath, fileBytes);
     }
 
-    private string GenerateTargetFilePath(Models.Roles.Targets.TargetMetadata targetFile) => Path.Combine(_config.LocalTargetsDir, targetFile.Path.RelPath);
-
-    private async Task<TUF.Models.Roles.Targets.TargetMetadata> preOrderDepthFirstWalk(string targetFilePath)
+    private void VerifyTargetFile(TargetFile targetFile, byte[] data)
     {
+        // Verify length
+        if (data.Length != targetFile.Length)
+        {
+            throw new Exception($"Target file length mismatch. Expected: {targetFile.Length}, Actual: {data.Length}");
+        }
+
+        // Verify at least one hash
+        bool validHash = false;
+        foreach (var hash in targetFile.Hashes)
+        {
+            var computedHash = hash.Key.ToLowerInvariant() switch
+            {
+                "sha256" => Convert.ToHexString(System.Security.Cryptography.SHA256.HashData(data)).ToLowerInvariant(),
+                "sha512" => Convert.ToHexString(System.Security.Cryptography.SHA512.HashData(data)).ToLowerInvariant(),
+                _ => null
+            };
+            
+            if (computedHash != null && computedHash == hash.Value.ToLowerInvariant())
+            {
+                validHash = true;
+                break;
+            }
+        }
+
+        if (!validHash)
+        {
+            throw new Exception("No valid hash found for target file");
+        }
+    }
+
+    private async Task<(string Path, TargetFile File)> PreOrderDepthFirstWalk(string targetFilePath)
+    {
+        if (_trusted is not CompleteTrustedMetadata complete)
+        {
+            throw new Exception("Complete trusted metadata not loaded");
+        }
+
         Stack<(string Role, string Parent)> toVisit = new();
-        toVisit.Push((Models.Roles.Targets.TargetsRole.TypeLabel, Models.Roles.Root.Root.TypeLabel));
+        toVisit.Push(("targets", "root"));
         HashSet<string> visited = new();
         while (toVisit.Count < _config.MaxDelegations && toVisit.Count > 0)
         {
@@ -139,17 +180,17 @@ public class Updater
                 continue;
             }
             await LoadTargets(role, parent);
-            CoreTrustedMetadata rtts = _trusted as CoreTrustedMetadata ?? throw new Exception("trusted metadata not loaded");
-            var targets = rtts.Targets[role]!;
-            if (targets.Signed.Targets.TryGetValue(new(targetFilePath), out var targetMeta))
+            CompleteTrustedMetadata rtts = _trusted as CompleteTrustedMetadata ?? throw new Exception("trusted metadata not loaded");
+            var targets = rtts.Targets[role];
+            if (targets is not null && targets.Signed.TargetMap.TryGetValue(targetFilePath, out var targetMeta))
             {
                 // if we have the file, return it
-                return targetMeta;
+                return (targetFilePath, targetMeta);
             }
             // if we didn't find the file in this role
             visited.Add(role); // add the visited role
             // add any child delegations to the list to traverse
-            if (targets.Signed.Delegations is Models.Roles.Targets.Delegations delegations)
+            if (targets is not null && targets.Signed.Delegations is {} delegations)
             {
                 var roles = delegations.GetRolesForTarget(targetFilePath);
                 foreach (var r in roles)
@@ -176,7 +217,7 @@ public class Updater
         await LoadRoot();
         await LoadTimestamp();
         await LoadSnapshot();
-        await LoadTargets(Models.Roles.Targets.TargetsRole.TypeLabel, Models.Roles.Root.Root.TypeLabel);
+        await LoadTargets("targets", "root");
     }
 
     private async Task LoadRoot()
@@ -188,7 +229,7 @@ public class Updater
         {
             try
             {
-                data = await DownloadMetadata(Models.Roles.Root.Root.TypeLabel, _config.RootMaxLength, v);
+                data = await DownloadMetadata("root", _config.RootMaxLength, v);
             }
             catch (HttpRequestException ex) when (ex.StatusCode == System.Net.HttpStatusCode.NotFound)
             {
@@ -198,27 +239,27 @@ public class Updater
         if (data is not null)
         {
             _trusted = _trusted.UpdateRoot(data);
-            await PersistMetadata(Models.Roles.Root.Root.TypeLabel, data);
+            await PersistMetadata("root", data);
         }
     }
 
     private async Task LoadTimestamp()
     {
-        var data = await DownloadMetadata(Models.Roles.Timestamp.Timestamp.TypeLabel, _config.TimestampMaxLength, null);
-        _trusted = _trusted.UpdateTimeStamp(data);
-        await PersistMetadata(Models.Roles.Timestamp.Timestamp.TypeLabel, data);
+        var data = await DownloadMetadata("timestamp", _config.TimestampMaxLength, null);
+        _trusted = _trusted.UpdateTimestamp(data);
+        await PersistMetadata("timestamp", data);
     }
 
     private async Task LoadSnapshot()
     {
-        if (_trusted is RootAndTimestampTrustedMetadata rts)
+        if (_trusted is TrustedMetadataWithTimestamp rts)
         {
-            var snapshotMeta = rts.Timestamp.Signed.SnapshotFileMetadata;
-            var length = snapshotMeta.Length ?? _config.SnapshotMaxLength;
+            var snapshotMeta = rts.Timestamp.Signed.Meta["snapshot.json"];
+            var length = (uint)(snapshotMeta.Length ?? _config.SnapshotMaxLength);
             int? version = rts.Root.Signed.ConsistentSnapshot is true ? (int)snapshotMeta.Version : null;
-            var data = await DownloadMetadata(Models.Roles.Snapshot.Snapshot.TypeLabel, _config.SnapshotMaxLength, version);
+            var data = await DownloadMetadata("snapshot", length, version);
             _trusted = _trusted.UpdateSnapshot(data, isTrusted: false);
-            await PersistMetadata(Models.Roles.Snapshot.Snapshot.TypeLabel, data);
+            await PersistMetadata("snapshot", data);
         }
         else {
             throw new Exception("trusted timestamp not loaded");
@@ -227,17 +268,17 @@ public class Updater
 
     private async Task LoadTargets(string roleName, string parentName)
     {
-        if (_trusted is RootAndTimestampAndSnapshotTrustedMetadata rts)
+        if (_trusted is TrustedMetadataWithSnapshot rts)
         {
-            if (!rts.Snapshot.Signed.Meta.TryGetValue(new(roleName + ".json"), out var roleFileMetadata))
+            if (!rts.Snapshot.Signed.Meta.TryGetValue($"{roleName}.json", out var roleFileMetadata))
             {
                 throw new Exception($"Role file metadata for {roleName} not found");
             }
 
             var length = roleFileMetadata.Length ?? _config.TargetsMaxLength;
             int? version = rts.Root.Signed.ConsistentSnapshot is true ? (int)roleFileMetadata.Version : null;
-            var data = await DownloadMetadata(roleName, length, version);
-            _trusted = _trusted.UpdateDelegatedTargets(data, roleName, parentName);
+            var data = await DownloadMetadata(roleName, (uint)length, version);
+            _trusted = rts.UpdateDelegatedTargets(data, roleName, parentName);
             await PersistMetadata(roleName, data);
         }
         else
